requirements.txt
affine==2.4.0
aiohttp==3.9.0b0
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.12.14
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
numpy==1.26.2
propcache==0.2.1
pyparsing==3.2.1
pyserial==3.5
rasterio==1.3.9
rasterioxyz==0.0.2
snuggs==1.4.7
websockets==13.1
yarl==1.18.3
